pyproject.toml
[tool.poetry]
name = "UTF-Tech Test Assignment"
version = "0.1.0"
description = "Тестовое задание на Django для UTF-Tech"
authors = ["Павел Костюк <pavkos0725@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.3"
django-rest-framework = "^0.1.0"
django-model-utils = "^5.0.0"
python-dotenv = "^1.0.1"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
